+++ flexbot.egg-info/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 1.0
-Name: slackbot-workout
-Version: 0.3
+Name: flexbot
+Version: 0.1
 Summary: A fun framework for incorporating more exercise into your life at work through Slack.
 Home-page: UNKNOWN
 Author: Brandon Shin, Miles Yucht
