+++ flexbot.egg-info/SOURCES.txt
@@ -1,0 +1,20 @@
+MANIFEST.in
+setup.py
+flexbot/__init__.py
+flexbot/api.py
+flexbot/bot.py
+flexbot/configurators.py
+flexbot/constants.py
+flexbot/exercise.py
+flexbot/logger_factory.py
+flexbot/loggers.py
+flexbot/manager.py
+flexbot/server.py
+flexbot/user.py
+flexbot/util.py
+flexbot/web.py
+flexbot.egg-info/PKG-INFO
+flexbot.egg-info/SOURCES.txt
+flexbot.egg-info/dependency_links.txt
+flexbot.egg-info/requires.txt
+flexbot.egg-info/top_level.txt
