+++ flexbot.egg-info/requires.txt
@@ -2,3 +2,4 @@ slacker>=0.8.6
 psycopg2>=2.6.1
 cherrypy>=4.0.0
 pyyaml>=3.11
+pystache>=0.5.4
