+++ flexbot.egg-info/top_level.txt
@@ -1,0 +1,3 @@
+test
+flexbot
+samples
