+++ flexbot/util.py
