--- slackbot_workout.egg-info/SOURCES.txt
@@ -1,16 +1,0 @@
-MANIFEST.in
-setup.py
-slackbot_workout/__init__.py
-slackbot_workout/api.py
-slackbot_workout/bot.py
-slackbot_workout/configurators.py
-slackbot_workout/loggers.py
-slackbot_workout/manager.py
-slackbot_workout/server.py
-slackbot_workout/user.py
-slackbot_workout/web.py
-slackbot_workout.egg-info/PKG-INFO
-slackbot_workout.egg-info/SOURCES.txt
-slackbot_workout.egg-info/dependency_links.txt
-slackbot_workout.egg-info/requires.txt
-slackbot_workout.egg-info/top_level.txt
