--- slackbot_workout.egg-info/top_level.txt
@@ -1,3 +1,0 @@
-samples
-slackbot_workout
-test
